# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Auth / UserManager",
    "version": "1.0.0",
    "description": "Documentation de l'API Auth / UserManager (SIWE, OTP, JWT, KYC)"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Serveur local"
    }
  ],
  "paths": {
...
    "/api/wallets": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Créer un nouveau wallet -> C'EST OK",
        "description": "Crée un nouveau wallet blockchain pour l'utilisateur spécifié.\n\n**Fonctionnalités :**\n- Génère automatiquement une adresse Ethereum\n- Chiffre la clé privée de manière sécurisée\n- Le premier wallet créé devient le wallet par défaut\n- Initialise le solde à 0 pour la devise choisie\n",
        "operationId": "createWallet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "label"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID de l'utilisateur",
                    "example": 1
                  },
                  "label": {
                    "type": "string",
                    "description": "Nom du wallet",
                    "example": "Mon Wallet Principal"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "PERSONAL",
                      "BUSINESS",
                      "SAVINGS"
                    ],
                    "description": "Type de wallet",
                    "default": "PERSONAL",
                    "example": "PERSONAL"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Devise initiale du wallet",
                    "default": "XAF",
                    "example": "XAF"
                  }
                }
              },
              "examples": {
                "personal": {
                  "summary": "Wallet personnel",
                  "value": {
                    "userId": 1,
                    "label": "Mon Wallet Principal",
                    "type": "PERSONAL",
                    "currency": "XAF"
                  }
                },
                "business": {
                  "summary": "Wallet professionnel",
                  "value": {
                    "userId": 1,
                    "label": "Entreprise SARL",
                    "type": "BUSINESS",
                    "currency": "EUR"
                  }
                },
                "savings": {
                  "summary": "Wallet épargne",
                  "value": {
                    "userId": 1,
                    "label": "Mon Épargne",
                    "type": "SAVINGS",
                    "currency": "USDT"
                  }
                },
                "minimal": {
                  "summary": "Création minimale (valeurs par défaut)",
                  "value": {
                    "userId": 1,
                    "label": "Mon Wallet"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Wallet créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "walletId": {
                          "type": "integer",
                          "example": 1
                        },
                        "address": {
                          "type": "string",
                          "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                        },
                        "label": {
                          "type": "string",
                          "example": "Mon Wallet Principal"
                        },
                        "type": {
                          "type": "string",
                          "example": "PERSONAL"
                        },
                        "isDefault": {
                          "type": "boolean",
                          "example": true
                        },
                        "status": {
                          "type": "string",
                          "example": "ACTIVE"
                        },
                        "message": {
                          "type": "string",
                          "example": "Wallet created successfully"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "missingFields": {
                    "summary": "Champs manquants",
                    "value": {
                      "error": "Missing required fields: userId, label"
                    }
                  },
                  "userNotFound": {
                    "summary": "Utilisateur non trouvé",
                    "value": {
                      "error": "user does not exist"
                    }
                  },
                  "invalidType": {
                    "summary": "Type invalide",
                    "value": {
                      "error": "Invalid wallet type"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Authentication required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Récupérer tous les wallets de l'utilisateur  -> C'EST OK",
        "description": "Retourne la liste complète des wallets appartenant à l'utilisateur authentifié.\n\n**Informations retournées :**\n- Détails de chaque wallet\n- Soldes pour toutes les devises\n- Wallet par défaut identifié\n- Triés par : wallet par défaut en premier, puis par date de création\n",
        "operationId": "getUserWallets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des wallets récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "description": "Nombre total de wallets",
                          "example": 3
                        },
                        "wallets": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Wallet"
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "multipleWallets": {
                    "summary": "Utilisateur avec plusieurs wallets",
                    "value": {
                      "success": true,
                      "data": {
                        "total": 3,
                        "wallets": [
                          {
                            "id": 1,
                            "address": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                            "label": "Mon Wallet Principal",
                            "type": "PERSONAL",
                            "status": "ACTIVE",
                            "isDefault": true,
                            "balances": [
                              {
                                "currency": "XAF",
                                "balance": "125000.50",
                                "currencyName": "CFA Franc",
                                "currencySymbol": "FCFA"
                              },
                              {
                                "currency": "EUR",
                                "balance": "500.00",
                                "currencyName": "Euro",
                                "currencySymbol": "€"
                              }
                            ],
                            "createdAt": "2025-01-10T10:00:00Z"
                          },
                          {
                            "id": 2,
                            "address": "0x8F3Cf7ad23Cd3CaDbD9735AFf958023239c6A063",
                            "label": "Mon Épargne",
                            "type": "SAVINGS",
                            "status": "ACTIVE",
                            "isDefault": false,
                            "balances": [
                              {
                                "currency": "USDT",
                                "balance": "1000.00",
                                "currencyName": "Tether USD",
                                "currencySymbol": "$"
                              }
                            ],
                            "createdAt": "2025-01-12T15:30:00Z"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/wallets/{walletId}": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Récupérer les détails d'un wallet spécifique  -> C'EST OK",
        "description": "Retourne les informations détaillées d'un wallet.\nL'utilisateur ne peut accéder qu'à ses propres wallets.\n",
        "operationId": "getWalletById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "required": true,
            "description": "ID du wallet",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du wallet récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Wallet"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID de wallet invalide"
          },
          "401": {
            "description": "Non authentifié"
          },
          "404": {
            "description": "Wallet non trouvé ou accès refusé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Wallet not found or access denied"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
...
    "/api/wallets/rate/{currency}": {
      "get": {
        "tags": [
          "Admin Wallets"
        ],
        "summary": "Récupérer le taux de change d'une devise",
        "description": "Retourne le taux de change configuré sur le contrat `MultiWalletExchange`.\n",
        "operationId": "getExchangeRate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "currency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EUR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Taux récupéré"
          },
          "404": {
            "description": "Devise inconnue"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Wallet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID unique du wallet",
            "example": 1
          },
          "address": {
            "type": "string",
            "description": "Adresse Ethereum du wallet",
            "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
          },
          "label": {
            "type": "string",
            "description": "Nom personnalisé du wallet",
            "example": "Mon Wallet Principal"
          },
          "type": {
            "type": "string",
            "enum": [
              "PERSONAL",
              "BUSINESS",
              "SAVINGS"
            ],
            "description": "Type de wallet",
            "example": "PERSONAL"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "FROZEN",
              "CLOSED"
            ],
            "description": "Statut du wallet",
            "example": "ACTIVE"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Wallet par défaut de l'utilisateur",
            "example": true
          },
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletBalance"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T10:30:00Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenu lors du login"
      }
    }
  },
...
}